+++ envgap/__init__.py
@@ -1,3 +1,3 @@
 """envgap finds gaps in Python environment config."""
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"